solver: move spec clause generation to its own module
This commit moves spec clauses generation in
solver/clauses.py, together with the version constraints,
target constraints and variant values it discovers
along the way, which the setup reads back once generation
is done.

The clauses and their order are unchanged: the ASP program
text is byte identical for abstract, namespaced and reused
specs.
